% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optics.R
\name{cluster_by_reachability}
\alias{cluster_by_reachability}
\title{Cluster by Reachability}
\usage{
cluster_by_reachability(data, threshold)
}
\arguments{
\item{data}{matrix; input vectors, with each column of the matrix being a vector. \code{nrow(data)} is the ector dimension.}

\item{threshold}{double threshold used to determine clusters}
}
\value{
returns \code{data} with an addition \code{"cluster"} attribute containing the clusters. Noise is marked as -1.
}
\description{
Extract cluster via a reachability threshold. Using the correct threshold extracts the DBSCAN clustering. This correct threshold might for example be guessed by looking at the reachability plot with each valley being a cluster. This naive approach is inferior to the extract_DBSCAN_clustering function!
}
\examples{
data <- generate_2d_cluster(200);
data <- OPTICS(data, .2, 25);
cluster_by_reachability(data, .15);
}
