% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering-package.R
\docType{package}
\name{clustering-package}
\alias{clustering}
\alias{clustering-package}
\title{clustering: R-Course Project 13: Clustering}
\description{
This package implements a variety of clustering algorithms and some related utility as part of the R-Course during the summer term of 2020.
}
\author{
\strong{Maintainer}: Ahmad Allah \email{Fadl@stud.uni-heidelberg.de} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

Authors:
\itemize{
  \item Mitodru Niyogi \email{Mitodru.Niyogi@stud.uni-heidelberg.de} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})
  \item Fabian Chrubasik \email{Chrubasik@stud.uni-heidelberg.de} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})
  \item Georg Weisert \email{Weisert@stud.uni-heidelberg.de} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})
}

}
\keyword{internal}
