% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_clustering.R
\name{create_training_data}
\alias{create_training_data}
\title{Create Training Data.}
\usage{
create_training_data(n)
}
\arguments{
\item{n}{integer; the number of vectors in the returned training data.}
}
\value{
returns a matrix with two rows and \code{n} columns; each column is a training vector.
}
\description{
This function creates training data as specified in "Beispiel 10.23" in reference "Richter19". It is meant for testing purposes.
}
