% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_clustering.R
\name{create_training_vector}
\alias{create_training_vector}
\title{Create a Single Training Vector}
\usage{
create_training_vector()
}
\value{
returns a single training vector
}
\description{
This function creates a single training vector as specified in "Beispiel 10.23" in reference "Richter19"
}
\references{
Richter, S. (2019). Statistisches und maschinelles Lernen. Springer Spektrum.
}
