% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_medoids.R
\name{dissim}
\alias{dissim}
\title{dissimialrity}
\usage{
dissim(data, k)
}
\arguments{
\item{data}{matrix; columns are vectors}

\item{k}{integer; column of which the dissimilarities are calculated}
}
\value{
vector of all dissimilarities
}
\description{
calculates all dissimilarities from a data point
}
\examples{
data <- matrix(c(1,1.1,1,1,2,2,2,2.1), ncol=4)
dissim(data, 3)
}
