% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_utility.R
\name{euclidean_distance}
\alias{euclidean_distance}
\title{Calculate the Euclidean distance between two vectors.}
\usage{
euclidean_distance(x, y)
}
\arguments{
\item{x}{a vector representing a (mathematical) vector}

\item{y}{a vector representing a (mathematical) vector}
}
\value{
double; Euclidean distance between \code{x} and \code{y}
}
\description{
\code{euclidean_distance} calculates the squared Euclidean norm of a vector.
}
\examples{
x = c(1,2,3)
y = c(2,3,4)
euclidean_distance(x, y)
# 1.732051
}
