% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_utility.R
\name{euclidean_norm}
\alias{euclidean_norm}
\title{Calculate the squared Euclidean Norm of a Vector.}
\usage{
euclidean_norm(x)
}
\arguments{
\item{x}{a vector representing a (mathematical) vector}
}
\value{
double; Euclidean norm of \code{x}
}
\description{
\code{euclidean_norm} calculates the Euclidean norm of a vector.
}
\examples{
x <- c(1,2,3)
euclidean_norm(x)
# 3.741657
}
