% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_utility.R
\name{euclidean_norm_squared}
\alias{euclidean_norm_squared}
\title{Calculate the squared Euclidean Norm of a Vector.}
\usage{
euclidean_norm_squared(x)
}
\arguments{
\item{x}{a vector representing a (mathematical) vector}
}
\value{
double; squared Euclidean norm of \code{x}
}
\description{
\code{euclidean_norm_squared} calculates the squared Euclidean norm of a vector.
}
\examples{
x = c(1,2,3)
euclidean_norm_squared(x)
# 14
}
