% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optics.R
\name{extract_DBSCAN_clustering}
\alias{extract_DBSCAN_clustering}
\title{Extract DBSCAN Clustering}
\usage{
extract_DBSCAN_clustering(data, eps_prime = NULL)
}
\arguments{
\item{data}{matrix; data processed by the OPTICS algorithm}

\item{eps_prime}{double; radius used for cluster extraction. \code{eps_prime} has to be less than or equal to the \code{eps} used in the OPTICS algorithm. If this is not set the OPTICS's \code{eps} value is used instead.}
}
\value{
The input data with the \code{"cluster"} attribute added which contains the clustering information.
}
\description{
Extract the DBSCAN clustering from data processed by the OPTICS algorithm according to the paper "OPTICS: Ordering Points To Identify the Clustering Structure".
}
\examples{
data <- matrix(c(1,1.1,1,1,2,2,2,2.1), ncol=4)
opt_data <- OPTICS(data, .2, 1)
extract_DBSCAN_clustering(opt_data, .2)
}
