% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_clustering.R
\name{gaussian_core}
\alias{gaussian_core}
\title{Calculate the Gaussian Core for Two Input Vectors}
\usage{
gaussian_core(x, y, gamma = 7.5)
}
\arguments{
\item{x}{vector}

\item{y}{vector}

\item{gamma}{double; has to be a positive number; defaults to \code{7.5}}
}
\value{
returns the gaussian code of \code{x} and \code{y} using \code{gamma}
}
\description{
\code{gaussian_core} calculates the gaussian core of two given vectors \code{x} and \code{y} of the same length using \code{gamma}.
}
