% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_utility.R
\name{gaussian_kernel}
\alias{gaussian_kernel}
\title{Calculate the Gaussian kernel for Two Input Vectors}
\usage{
gaussian_kernel(x, y, gamma = 7.5)
}
\arguments{
\item{x}{vector}

\item{y}{vector}

\item{gamma}{double; has to be a positive number; defaults to \code{7.5}}
}
\value{
returns the gaussian kernel of \code{x} and \code{y} using \code{gamma}
}
\description{
\code{gaussian_core} calculates the gaussian kernel of two given vectors \code{x} and \code{y} of the same length using \code{gamma}.
}
