% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_utility.R
\name{gaussian_kernel_with_fixed_gamma}
\alias{gaussian_kernel_with_fixed_gamma}
\title{Gaussian Kernel factory}
\usage{
gaussian_kernel_with_fixed_gamma(gamma)
}
\arguments{
\item{gamma}{The \code{gamma} value to be used for the Gaussien Kernel.}
}
\value{
A function that calculates the Gaussian Kernel for input vector \code{x}, \code{y} with a fixed \code{gamma}.
}
\description{
Returns a function that calculates the Gaussian Kernel with \code{gamma} fixed to the passed value.
}
\examples{
f <- gaussian_kernel_with_fixed_gamma(3)
}
