% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{generate_2d_cluster}
\alias{generate_2d_cluster}
\title{Generate a Cluster of Two Dimensional Vectors}
\usage{
generate_2d_cluster(n, center = c(0, 0))
}
\arguments{
\item{n}{integer; number of vectors that comprise the returned cluster}

\item{center}{vector; vector of two elements specifying thw x,y coordinates of the center point}
}
\value{
Returns a matrix with two rows and n columns containing the generated cluster of two dimensional vectors as column-vectors.
}
\description{
Generates a cluster of two dimensional vectors. The number of vectors and the center of the cluster may be specified via the function's arguments.
}
\examples{
aCluster <- generate_2d_cluster(100);
anotherCluster <- generate_2d_cluster(200, center=c(3,2));
}
