% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{generate_nested_2d_training_data}
\alias{generate_nested_2d_training_data}
\title{Generate Nested 2D Training Data.}
\usage{
generate_nested_2d_training_data(n)
}
\arguments{
\item{n}{integer; the number of vectors in the returned training data.}
}
\value{
returns a matrix with two rows and \code{n} columns; each column is a training vector.

returns a single training vector
}
\description{
This function generates 2D training data as specified in "Beispiel 10.23" in reference "Richter19". The result will be an inner clusters surrounded by a ring of points, the outer cluster. This data is especially useful to test spectral clustering.

This function generates a single training vector as specified in "Beispiel 10.23" in reference "Richter19"
}
\examples{
generate_training_data(10)
Generate a Single Training Vector
}
\references{
Richter, S. (2019). Statistisches und maschinelles Lernen. Springer Spektrum.
}
