% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{generate_training_data}
\alias{generate_training_data}
\title{Generate Training Data.}
\usage{
generate_training_data(n)
}
\arguments{
\item{n}{integer; the number of vectors in the returned training data.}
}
\value{
returns a matrix with two rows and \code{n} columns; each column is a training vector.
}
\description{
This function generates training data as specified in "Beispiel 10.23" in reference "Richter19". It is meant for testing purposes.
}
