% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{generate_training_vector}
\alias{generate_training_vector}
\title{Generate a Single Training Vector}
\usage{
generate_training_vector()
}
\value{
returns a single training vector
}
\description{
This function generates a single training vector as specified in "Beispiel 10.23" in reference "Richter19"
}
\references{
Richter, S. (2019). Statistisches und maschinelles Lernen. Springer Spektrum.
}
