% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_clustering.R
\name{hclust_order}
\alias{hclust_order}
\title{Hieraerchical clusters order}
\usage{
hclust_order(mat)
}
\arguments{
\item{mat}{matrix; an n−1 by 2 matrix}
}
\value{
vector in ordered manner given the permutation of the original observations for plotting
}
\description{
The function returns a vector giving the permutation of the original
}
\examples{
data <- matrix(c(1,1.1,1,1,2,2,2,2.1), ncol=4)
hclust_order(distance(data))
}
