% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_clustering.R
\name{hierclust}
\alias{hierclust}
\title{Hierarchical clustering implementation. It can bee used like original hclust(d, method).
Hierarchical Clustering over a Matrix's Column-Vectors.}
\usage{
hierclust(d, method = c("single", "complete", "average", "median"))
}
\arguments{
\item{d}{matrix; input matrix columns are vectors}

\item{method}{string; linkage method usage}
}
\value{
the list structure object(labels, call, method, order, height, merge, dist.method) as similar to the output from hclust.
}
\description{
Performs the agglomerative hierarchical clustering algorithm describe in "Definition 10.3.2" in reference James19
}
\examples{
data <- distance(USArrests)
data_clust <- hierclust(data, method="single")
}
\references{
Gareth James, Daniela Witten, Trevor Hastie, Robert Tibshirani. (2013). An Introduction to Statistical Learning with Applications in R.
}
