% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_utility.R
\name{induced_metric_factory}
\alias{induced_metric_factory}
\title{Induced Metric Factory}
\usage{
induced_metric_factory(vector_norm)
}
\arguments{
\item{vector_norm}{function; a vector norm}
}
\value{
function; calculates the metric induced by \code{vector_norm}.
}
\description{
This function returns a function calculating the metric induced by the given vector norm.
}
\examples{
euclidean_metric <- induced_metric_factory(euclidean_norm)
x = c(1,2,3)
y = c(2,3,4)
euclidean_metric(x, y)
# 1.732051
}
