% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_utility.R
\name{infinity_norm}
\alias{infinity_norm}
\title{Calculate the Infinity-Norm of a Vector.}
\usage{
infinity_norm(x)
}
\arguments{
\item{x}{a vector representing a (mathematical) vector}
}
\value{
double; infinity-norm of \code{x}
}
\description{
\code{infinity_norm} calculates the infinity-norm of a vector. Also known as maximum norm.
}
\examples{
x = c(1,2,3)
infinity_norm(x)
# 3
}
