% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_medoids.R
\name{k_medoids}
\alias{k_medoids}
\title{k-medoids algorithm}
\usage{
k_medoids(data, k = 2)
}
\arguments{
\item{data}{matrix; columns are vectors}

\item{k}{integer; number of clusters}
}
\value{
input data with attribute "cluster" assigning a cluster to every column-vector.
}
\description{
Finds clusters in the input data using the k_medoids algorithm.
}
\examples{
data <- matrix(c(1,1.1,1,1,2,2,2,2.1), ncol=4)
k_medoids(data, 2)
}
