% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeanalgorithm.R
\name{m}
\alias{m}
\title{Chose random centerns for k clusters}
\usage{
m(mydata, k)
}
\arguments{
\item{k}{double; number of cluster}

\item{data}{matrix; matrix columns are (mathematical) vectors. Input data.}
}
\value{
matrix; with k columns .
}
\description{
chose random centers for cluster   it will only be used internely in kmean clustering algorithm.
}
\examples{
data <- matrix(c(1,1.1,1,1,2,2,2,2.1), ncol=4)
m(data, 3)
}
