% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_clustering.R
\name{norm_2}
\alias{norm_2}
\title{Calculate the squared Euclidian Norm of a Vector.}
\usage{
norm_2(x)
}
\arguments{
\item{x}{a vector representing a vector}
}
\value{
double; euclidian norm of \code{x}
#' @export
}
\description{
\code{norm_2} calculates the euclidian norm of a vector.
}
\examples{
x = c(1,2,3)
norm_2(x)
# 3.741657
}
