% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_clustering.R
\name{norm_2_sq}
\alias{norm_2_sq}
\title{Calculate the squared Euclidian Norm of a Vector.
@description \code{norm_2_sq} calculates the squared euclidian norm of a vector.}
\usage{
norm_2_sq(x)
}
\arguments{
\item{x}{a vector representing a vector}
}
\value{
double; squared euclidian norm of \code{x}
}
\description{
Calculate the squared Euclidian Norm of a Vector.
@description \code{norm_2_sq} calculates the squared euclidian norm of a vector.
}
\examples{
x = c(1,2,3)
norm_2_sq(x)
# 14
}
