% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_utility.R
\name{one_norm}
\alias{one_norm}
\title{Calculate the 1-Norm of a Vector.}
\usage{
one_norm(x)
}
\arguments{
\item{x}{a vector representing a (mathematical) vector}
}
\value{
double; 1-norm of \code{x}
}
\description{
\code{one_norm} calculates the Euclidean norm of a vector. Also known as Taxicab norm or Manhattan norm.
}
\examples{
x = c(1,2,3)
one_norm(x)
# 6
}
