% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_utility.R
\name{p_norm}
\alias{p_norm}
\title{Calculate the p-Norm of a Vector.}
\usage{
p_norm(x, p)
}
\arguments{
\item{x}{a vector representing a (mathematical) vector}

\item{p}{integer; the p value}
}
\value{
double; p-norm of \code{x}
}
\description{
\code{p_norm} calculates the p-norm of a vector for a given \code{p}.
}
\examples{
x <- c(1,2,3)
p_norm(x, 2)
# 3.741657
}
