% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_utility.R
\name{p_norm_factory}
\alias{p_norm_factory}
\title{p-Norm Function Factory}
\usage{
p_norm_factory(p)
}
\arguments{
\item{p}{integer; specify which p-norm the returned function will compute.}
}
\value{
function; A function which computes the p-norm for the given \code{p}.
}
\description{
Generate a function calculating the p-norm for the given p.
}
\examples{
two_norm <- p_norm_factory(2)
two_norm(c(1,2,3))
# 3.741657
}
