% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{plot_clustered_2d_data}
\alias{plot_clustered_2d_data}
\title{Plot Clustered 2D Data}
\usage{
plot_clustered_2d_data(
  data,
  point_size = 0.75,
  show_noise = TRUE,
  show_legend = FALSE,
  hide_axis_text = FALSE
)
}
\arguments{
\item{data}{matrix; columns are vectors. Has to have the \code{"cluster"} attribute set.}

\item{point_size}{numeric; size of the points}

\item{show_noise}{logical; toggle plotting of noise}

\item{show_legend}{logical; toggle legend display}
}
\value{

}
\description{
Plot Clustered 2D Data
}
\examples{
data <- matrix(c(1,1,2,1));
attr(data, "cluster") <- c(1,2);
plot_clustered_2d_data(data);
}
