% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optics.R
\name{reachability_plot}
\alias{reachability_plot}
\title{Reachabilty Plot}
\usage{
reachability_plot(data, threshold = NULL)
}
\arguments{
\item{data}{matrix; data has to be the result of a previous run of the OPTICS algorithm.}

\item{threshold}{numeric; show the given threshold in the plot, optinal.}
}
\description{
Generates a reachability plot for data resulting from the OPTICS function.
}
\examples{
data <- generate_2d_cluster(200);
data <- OPTICS(data, .2, 25);
reachability_plot(data, .15);
}
