% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optics.R
\name{OPTICS}
\alias{OPTICS}
\title{Ordering points to identify the clustering structure (OPTICS)}
\usage{
OPTICS(data, eps, minPts, distanceFunction = euclidean_distance)
}
\arguments{
\item{data}{matrix; matrix columns are (mathematical) vectors. Input data.}

\item{eps}{double; neighborhood radius}

\item{distanceFunction}{function; function calculating the distance between two vectors of length \code{nrow(data)}}

\item{MinPts}{integer; minimum number of points in neighborhood required}
}
\value{
matrix; input data with attribute "cluster" assigning a cluster to every column-vector.
}
\description{
Finds clusters in the input data using the OPTICS algorithm.
}
\examples{
data <- matrix(c(1,1.1,1,1,2,2,2,2.1), ncol=4)
DBSCAN(data, .2, 1)
}
